+++ Source/Liath.BigSpace.Domain/Jobs/Job.cs
@@ -8,7 +8,7 @@ namespace Liath.BigSpace.DataAccess.Definitions.Jobs
 {
     public abstract class Job
     {
-        public int JobID { get; set; }
+        public Int64 JobID { get; set; }
         public DateTime StartTime { get; set; }
         public TimeSpan Duration { get; set; }
     }
